Disabling the Fog on Farm_Objective Marker
So. I don't know if it's intentional, but currently the Objective Makers
are overlayed by fog, if an fog is set on the map.

since you call ignoreZ on it, so it renders throught the world, i
thought this was a bug.

[Currently](http://i.imgur.com/D1D3771.jpg)
[excepted](http://i.imgur.com/sJ5Kg9A.jpg)

This way will disabling the fog completly on the marker, so it will get
shown throught.